Move MemorizeMessage to contract.
